rollback to ns.3.29

+++ src/lte/helper/radio-environment-map-helper.h
@@ -61,13 +61,13 @@ public:
   /** 
    * \return the bandwidth (in num of RBs) over which SINR is calculated
    */
-  uint16_t GetBandwidth () const;
+  uint8_t GetBandwidth () const;
 
   /** 
    *
    * \param bw  the bandwidth (in num of RBs) over which SINR is calculated
    */
-  void SetBandwidth (uint16_t bw);
+  void SetBandwidth (uint8_t bw);
 
   /** 
    * Deploy the RemSpectrumPhy objects that generate the map according to the specified settings.
